+++ api-server/scanner-lib/src/sync/tests/mod.rs
@@ -13,6 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod simulation;
+
 use crate::blockchain_state::BlockchainState;
 
 use mempool::FeeRate;
@@ -21,9 +23,7 @@ use serialization::Encode;
 use super::*;
 
 use std::{
-    collections::{BTreeMap, BTreeSet},
     convert::Infallible,
-    num::NonZeroU64,
     sync::{Arc, Mutex},
     time::Duration,
 };
@@ -32,7 +32,6 @@ use api_server_common::storage::{
     impls::in_memory::transactional::TransactionalApiServerInMemoryStorage,
     storage_api::{
         ApiServerStorageRead, ApiServerStorageWrite, ApiServerTransactionRw, Transactional,
-        UtxoLock,
     },
 };
 
@@ -41,7 +40,6 @@ use chainstate_test_framework::{TestFramework, TransactionBuilder};
 use common::{
     address::Address,
     chain::{
-        block::timestamp::BlockTimestamp,
         output_value::OutputValue,
         signature::{
             inputsig::{
@@ -52,15 +50,10 @@ use common::{
         },
         stakelock::StakePoolData,
         timelock::OutputTimeLock,
-        tokens::{
-            make_token_id, NftIssuance, RPCIsTokenFrozen, RPCTokenInfo, RPCTokenTotalSupply,
-            TokenId,
-        },
-        CoinUnit, ConsensusUpgrade, DelegationId, Destination, GenBlockId, NetUpgrades,
-        OutPointSourceId, PoSChainConfigBuilder, PoolId, SignedTransaction, TxInput, TxOutput,
+        CoinUnit, Destination, OutPointSourceId, PoolId, SignedTransaction, TxInput, TxOutput,
         UtxoOutPoint,
     },
-    primitives::{per_thousand::PerThousand, Amount, BlockCount, CoinOrTokenId, Idable, H256},
+    primitives::{per_thousand::PerThousand, Amount, CoinOrTokenId, Idable, H256},
 };
 use crypto::{
     key::{KeyKind, PrivateKey},
@@ -1214,533 +1207,5 @@ async fn check_all_destinations_are_tracked(#[case] seed: Seed) {
         let utxos = db_tx.get_address_all_utxos(address.as_str()).await.unwrap();
         // check we have 2 utxos one locked and one unlocked
         assert_eq!(utxos.len(), 2);
-    }
-}
-
-#[rstest]
-#[trace]
-#[case(test_utils::random::Seed::from_entropy(), 20, 50)]
-#[tokio::test]
-async fn simulation(
-    #[case] seed: Seed,
-    #[case] max_blocks: usize,
-    #[case] max_tx_per_block: usize,
-) {
-    logging::init_logging();
-    let mut rng = make_seedable_rng(seed);
-
-    let (vrf_sk, vrf_pk) = VRFPrivateKey::new_from_rng(&mut rng, VRFKeyKind::Schnorrkel);
-    let (staking_sk, staking_pk) = PrivateKey::new_from_rng(&mut rng, KeyKind::Secp256k1Schnorr);
-    let (config_builder, genesis_pool_id) =
-        chainstate_test_framework::create_chain_config_with_default_staking_pool(
-            &mut rng, staking_pk, vrf_pk,
-        );
-
-    let upgrades = vec![(
-        BlockHeight::new(0),
-        ConsensusUpgrade::PoS {
-            initial_difficulty: None,
-            config: PoSChainConfigBuilder::new_for_unit_test()
-                .staking_pool_spend_maturity_block_count(BlockCount::new(5))
-                .build(),
-        },
-    )];
-    let consensus_upgrades = NetUpgrades::initialize(upgrades).expect("valid net-upgrades");
-
-    let epoch_length = NonZeroU64::new(rng.gen_range(1..10)).unwrap();
-    let sealed_epoch_distance_from_tip = rng.gen_range(1..10);
-    let chain_config = config_builder
-        .consensus_upgrades(consensus_upgrades)
-        .max_future_block_time_offset(std::time::Duration::from_secs(1_000_000))
-        .epoch_length(epoch_length)
-        .sealed_epoch_distance_from_tip(sealed_epoch_distance_from_tip)
-        .build();
-    let target_time = chain_config.target_block_spacing();
-    let genesis_pool_outpoint = UtxoOutPoint::new(chain_config.genesis_block_id().into(), 1);
-
-    // Initialize original TestFramework
-    let mut tf = TestFramework::builder(&mut rng)
-        .with_chain_config(chain_config.clone())
-        .with_initial_time_since_genesis(target_time.as_secs())
-        .with_staking_pools(BTreeMap::from_iter([(
-            genesis_pool_id,
-            (
-                staking_sk.clone(),
-                vrf_sk.clone(),
-                genesis_pool_outpoint.clone(),
-            ),
-        )]))
-        .build();
-
-    let storage = {
-        let mut storage = TransactionalApiServerInMemoryStorage::new(&chain_config);
-
-        let mut db_tx = storage.transaction_rw().await.unwrap();
-        db_tx.reinitialize_storage(&chain_config).await.unwrap();
-        db_tx.commit().await.unwrap();
-
-        storage
-    };
-    let mut local_state = BlockchainState::new(Arc::new(chain_config.clone()), storage);
-    local_state.scan_genesis(chain_config.genesis_block().as_ref()).await.unwrap();
-
-    let num_blocks = rng.gen_range((max_blocks / 2)..max_blocks);
-
-    let mut utxo_outpoints = Vec::new();
-    let mut staking_pools: BTreeSet<PoolId> = BTreeSet::new();
-    staking_pools.insert(genesis_pool_id);
-
-    let mut delegations = BTreeSet::new();
-    let mut token_ids = BTreeSet::new();
-
-    let mut data_per_block_height = BTreeMap::new();
-    data_per_block_height.insert(
-        BlockHeight::zero(),
-        (
-            utxo_outpoints.clone(),
-            staking_pools.clone(),
-            delegations.clone(),
-            token_ids.clone(),
-        ),
-    );
-
-    let mut tf_internal_staking_pools = BTreeMap::new();
-    tf_internal_staking_pools.insert(BlockHeight::zero(), tf.staking_pools.clone());
-
-    // Generate a random chain
-    for current_height in 0..num_blocks {
-        let create_reorg = rng.gen_bool(0.1);
-        let height_to_continue_from = if create_reorg {
-            rng.gen_range(0..=current_height)
-        } else {
-            current_height
-        };
-
-        tf = if create_reorg {
-            let blocks = if height_to_continue_from > 0 {
-                tf.chainstate
-                    .get_mainchain_blocks(BlockHeight::new(1), height_to_continue_from)
-                    .unwrap()
-            } else {
-                vec![]
-            };
-            let mut new_tf = TestFramework::builder(&mut rng)
-                .with_chain_config(chain_config.clone())
-                .with_initial_time_since_genesis(target_time.as_secs())
-                .with_staking_pools(BTreeMap::from_iter([(
-                    genesis_pool_id,
-                    (
-                        staking_sk.clone(),
-                        vrf_sk.clone(),
-                        genesis_pool_outpoint.clone(),
-                    ),
-                )]))
-                .build();
-            for block in blocks {
-                new_tf.progress_time_seconds_since_epoch(target_time.as_secs());
-                new_tf.process_block(block.clone(), BlockSource::Local).unwrap();
-            }
-            new_tf.staking_pools = tf_internal_staking_pools
-                .get(&BlockHeight::new(height_to_continue_from as u64))
-                .unwrap()
-                .clone();
-            new_tf.key_manager = tf.key_manager;
-
-            (utxo_outpoints, staking_pools, delegations, token_ids) = data_per_block_height
-                .get(&BlockHeight::new(height_to_continue_from as u64))
-                .unwrap()
-                .clone();
-
-            new_tf
-        } else {
-            tf
-        };
-
-        let block_height_to_continue_from = BlockHeight::new(height_to_continue_from as u64);
-        let mut prev_block_hash = tf
-            .chainstate
-            .get_block_id_from_height(&block_height_to_continue_from)
-            .unwrap()
-            .unwrap();
-
-        for block_height_idx in 0..=(current_height - height_to_continue_from) {
-            let block_height =
-                BlockHeight::new((height_to_continue_from + block_height_idx) as u64);
-
-            let mut block_builder = tf.make_pos_block_builder().with_random_staking_pool(&mut rng);
-
-            for _ in 0..rng.gen_range(10..max_tx_per_block) {
-                block_builder = block_builder.add_test_transaction(&mut rng);
-            }
-
-            let block = block_builder.build();
-            for tx in block.transactions() {
-                let new_utxos = (0..tx.inputs().len()).map(|output_index| {
-                    UtxoOutPoint::new(
-                        OutPointSourceId::Transaction(tx.transaction().get_id()),
-                        output_index as u32,
-                    )
-                });
-                utxo_outpoints.extend(new_utxos);
-
-                let new_pools = tx.outputs().iter().filter_map(|out| match out {
-                    TxOutput::CreateStakePool(pool_id, _) => Some(pool_id),
-                    TxOutput::Burn(_)
-                    | TxOutput::Transfer(_, _)
-                    | TxOutput::LockThenTransfer(_, _, _)
-                    | TxOutput::DataDeposit(_)
-                    | TxOutput::DelegateStaking(_, _)
-                    | TxOutput::CreateDelegationId(_, _)
-                    | TxOutput::IssueFungibleToken(_)
-                    | TxOutput::ProduceBlockFromStake(_, _)
-                    | TxOutput::IssueNft(_, _, _) => None,
-                });
-                staking_pools.extend(new_pools);
-
-                let new_delegations = tx.outputs().iter().filter_map(|out| match out {
-                    TxOutput::CreateDelegationId(_, _) => {
-                        let input0_outpoint =
-                            tx.inputs().iter().find_map(|input| input.utxo_outpoint()).unwrap();
-                        Some(make_delegation_id(input0_outpoint))
-                    }
-                    TxOutput::CreateStakePool(_, _)
-                    | TxOutput::Burn(_)
-                    | TxOutput::Transfer(_, _)
-                    | TxOutput::LockThenTransfer(_, _, _)
-                    | TxOutput::DataDeposit(_)
-                    | TxOutput::DelegateStaking(_, _)
-                    | TxOutput::IssueFungibleToken(_)
-                    | TxOutput::ProduceBlockFromStake(_, _)
-                    | TxOutput::IssueNft(_, _, _) => None,
-                });
-                delegations.extend(new_delegations);
-
-                let new_tokens = tx.outputs().iter().filter_map(|out| match out {
-                    TxOutput::IssueNft(_, _, _) | TxOutput::IssueFungibleToken(_) => {
-                        Some(make_token_id(tx.inputs()).unwrap())
-                    }
-                    TxOutput::CreateStakePool(_, _)
-                    | TxOutput::Burn(_)
-                    | TxOutput::Transfer(_, _)
-                    | TxOutput::LockThenTransfer(_, _, _)
-                    | TxOutput::DataDeposit(_)
-                    | TxOutput::DelegateStaking(_, _)
-                    | TxOutput::CreateDelegationId(_, _)
-                    | TxOutput::ProduceBlockFromStake(_, _) => None,
-                });
-                token_ids.extend(new_tokens);
-            }
-
-            prev_block_hash = block.get_id().into();
-            tf.process_block(block.clone(), BlockSource::Local).unwrap();
-
-            // save current state
-            tf_internal_staking_pools.insert(block_height.next_height(), tf.staking_pools.clone());
-            data_per_block_height.insert(
-                block_height.next_height(),
-                (
-                    utxo_outpoints.clone(),
-                    staking_pools.clone(),
-                    delegations.clone(),
-                    token_ids.clone(),
-                ),
-            );
-
-            local_state.scan_blocks(block_height, vec![block]).await.unwrap();
-        }
-
-        let block_height = BlockHeight::new(current_height as u64);
-        let median_time = tf.chainstate.calculate_median_time_past(&prev_block_hash).unwrap();
-
-        check_utxos(
-            &tf,
-            &local_state,
-            &utxo_outpoints,
-            median_time,
-            block_height.next_height(),
-        )
-        .await;
-
-        check_staking_pools(&tf, &local_state, &staking_pools).await;
-        check_delegations(&tf, &local_state, &delegations).await;
-        check_tokens(&tf, &local_state, &token_ids).await;
-    }
-}
-
-async fn check_utxos(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    utxos: &Vec<UtxoOutPoint>,
-    current_median_time: BlockTimestamp,
-    current_block_height: BlockHeight,
-) {
-    for outpoint in utxos {
-        check_utxo(
-            tf,
-            local_state,
-            outpoint.clone(),
-            current_median_time,
-            current_block_height,
-        )
-        .await;
-    }
-}
-
-async fn check_utxo(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    outpoint: UtxoOutPoint,
-    current_median_time: BlockTimestamp,
-    current_block_height: BlockHeight,
-) {
-    let c_utxo = tf.chainstate.utxo(&outpoint).unwrap();
-
-    // if this is a locked utxo get the unlock time/height
-    let unlock = c_utxo
-        .as_ref()
-        .and_then(|utxo| utxo.output().timelock().zip(utxo.source().blockchain_height().ok()))
-        .map(|(lock, height)| {
-            let block_id = tf.block_id(height.into_int());
-            let utxo_block_id = block_id.classify(tf.chainstate.get_chain_config());
-            let time_of_tx = match utxo_block_id {
-                GenBlockId::Block(id) => {
-                    tf.chainstate.get_block_header(id).unwrap().unwrap().timestamp()
-                }
-                GenBlockId::Genesis(_) => {
-                    tf.chainstate.get_chain_config().genesis_block().timestamp()
-                }
-            };
-            UtxoLock::from_output_lock(*lock, time_of_tx, height).into_time_and_height()
-        });
-
-    let tx = local_state.storage().transaction_ro().await.unwrap();
-
-    // fetch the locked utxo
-    let l_utxo = if let Some((unlock_time, unlock_height)) = unlock {
-        tx.get_locked_utxos_until_now(
-            unlock_height.unwrap_or(current_block_height),
-            (
-                current_median_time,
-                unlock_time.unwrap_or(current_median_time),
-            ),
-        )
-        .await
-        .unwrap()
-        .into_iter()
-        .find_map(|(out, info)| (out == outpoint).then_some(info))
-    } else {
-        None
-    };
-
-    // fetch the unlocked utxo
-    let s_utxo = tx.get_utxo(outpoint).await.unwrap();
-
-    match (c_utxo, s_utxo) {
-        (Some(c_utxo), Some(s_utxo)) => {
-            // if utxo is in chainstate it should not be spent
-            assert!(!s_utxo.spent());
-            // check outputs are the same
-            assert_eq!(c_utxo.output(), s_utxo.output());
-        }
-        (None, Some(s_utxo)) => {
-            // if utxo is not found in chainstate but found in scanner it must be spent
-            assert!(s_utxo.spent());
-            // and not in locked utxos
-            assert_eq!(l_utxo, None);
-        }
-        (Some(c_utxo), None) => {
-            // if utxo is in chainstate but not in unlocked scanner utxos it must be in the locked
-            // ones
-            if let Some(l_utxo) = l_utxo {
-                assert_eq!(c_utxo.output(), &l_utxo.output);
-            } else {
-                panic!("Utxo in chainstate but not in the scanner state");
-            }
-        }
-        (None, None) => {
-            // on reorg utxos will be gone from both chainstate and the scanner
-            // same for locked
-            assert_eq!(l_utxo, None);
-        }
-    };
-}
-
-async fn check_staking_pools(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    staking_pools: &BTreeSet<PoolId>,
-) {
-    for pool_id in staking_pools {
-        check_pool(tf, local_state, *pool_id).await;
-    }
-}
-
-async fn check_pool(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    pool_id: PoolId,
-) {
-    let tx = local_state.storage().transaction_ro().await.unwrap();
-    let scanner_data = tx.get_pool_data(pool_id).await.unwrap().unwrap();
-
-    if let Some(node_data) = tf.chainstate.get_stake_pool_data(pool_id).unwrap() {
-        // check staker total balances are the same
-        assert_eq!(node_data.staker_balance(), scanner_data.staker_balance());
-        // check the pledges are the same
-        assert_eq!(node_data.pledge_amount(), scanner_data.pledge_amount());
-        assert_eq!(node_data, scanner_data);
-
-        // Compare the delegation shares
-        let node_delegations = tf
-            .chainstate
-            .get_stake_pool_delegations_shares(pool_id)
-            .unwrap()
-            .unwrap_or_default();
-
-        let scanner_delegations = tx.get_pool_delegations(pool_id).await.unwrap();
-
-        // check all delegations from the node are contained in the scanner
-        for (id, share) in &node_delegations {
-            let scanner_delegation = scanner_delegations.get(id).unwrap();
-            // check the shares are the same
-            assert_eq!(share, scanner_delegation.balance());
-        }
-
-        // delegations that have not been staked yet are stored in the scanner but not in the node
-        for (id, scanner_delegation) in scanner_delegations {
-            let share = node_delegations.get(&id).unwrap_or(&Amount::ZERO);
-            // check the shares are the same
-            assert_eq!(share, scanner_delegation.balance());
-        }
-    } else {
-        // the pool has been decommissioned
-        assert_eq!(Amount::ZERO, scanner_data.pledge_amount());
-    }
-}
-
-async fn check_delegations(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    delegations: &BTreeSet<DelegationId>,
-) {
-    for delegation_id in delegations {
-        check_delegation(tf, local_state, *delegation_id).await
-    }
-}
-
-async fn check_delegation(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    delegation_id: DelegationId,
-) {
-    let tx = local_state.storage().transaction_ro().await.unwrap();
-    let scanner_data = tx.get_delegation(delegation_id).await.unwrap().unwrap();
-
-    if let Some(node_data) = tf.chainstate.get_stake_delegation_data(delegation_id).unwrap() {
-        assert_eq!(node_data.source_pool(), scanner_data.pool_id());
-        assert_eq!(
-            node_data.spend_destination(),
-            scanner_data.spend_destination()
-        );
-
-        // check delegation balances are the same
-        let node_delegation_balance = tf
-            .chainstate
-            .get_stake_delegation_balance(delegation_id)
-            .unwrap()
-            .unwrap_or(Amount::ZERO);
-        assert_eq!(node_delegation_balance, *scanner_data.balance());
-    } else {
-        // the pool has been decommissioned
-        assert_eq!(Amount::ZERO, *scanner_data.balance());
-    }
-}
-
-async fn check_tokens(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    token_ids: &BTreeSet<TokenId>,
-) {
-    for token_id in token_ids {
-        check_token(tf, local_state, *token_id).await;
-    }
-}
-
-async fn check_token(
-    tf: &TestFramework,
-    local_state: &BlockchainState<TransactionalApiServerInMemoryStorage>,
-    token_id: TokenId,
-) {
-    let tx = local_state.storage().transaction_ro().await.unwrap();
-    let node_data = tf.chainstate.get_token_info_for_rpc(token_id).unwrap().unwrap();
-
-    match node_data {
-        RPCTokenInfo::FungibleToken(node_data) => {
-            let scanner_data = tx.get_fungible_token_issuance(token_id).await.unwrap().unwrap();
-
-            assert_eq!(node_data.authority, scanner_data.authority);
-            assert_eq!(
-                node_data.token_ticker.into_bytes(),
-                scanner_data.token_ticker
-            );
-            assert_eq!(
-                node_data.metadata_uri.into_bytes(),
-                scanner_data.metadata_uri
-            );
-            assert_eq!(
-                node_data.number_of_decimals,
-                scanner_data.number_of_decimals
-            );
-            assert_eq!(
-                node_data.circulating_supply,
-                scanner_data.circulating_supply
-            );
-            assert_eq!(node_data.is_locked, scanner_data.is_locked);
-            assert_eq!(node_data.frozen, RPCIsTokenFrozen::new(scanner_data.frozen));
-            assert_eq!(
-                node_data.total_supply,
-                RPCTokenTotalSupply::from(scanner_data.total_supply)
-            );
-        }
-        RPCTokenInfo::NonFungibleToken(node_data) => {
-            let scanner_data = tx.get_nft_token_issuance(token_id).await.unwrap().unwrap();
-
-            match scanner_data {
-                NftIssuance::V0(scanner_data) => {
-                    assert_eq!(
-                        node_data.metadata.name.into_bytes(),
-                        scanner_data.metadata.name
-                    );
-                    assert_eq!(
-                        node_data.metadata.description.into_bytes(),
-                        scanner_data.metadata.description
-                    );
-                    assert_eq!(
-                        node_data.metadata.ticker.into_bytes(),
-                        scanner_data.metadata.ticker
-                    );
-                    assert_eq!(
-                        &node_data.metadata.icon_uri.map(|x| x.into_bytes()),
-                        scanner_data.metadata.icon_uri.as_ref()
-                    );
-                    assert_eq!(
-                        &node_data.metadata.additional_metadata_uri.map(|x| x.into_bytes()),
-                        scanner_data.metadata.additional_metadata_uri.as_ref()
-                    );
-                    assert_eq!(
-                        &node_data.metadata.media_uri.map(|x| x.into_bytes()),
-                        scanner_data.metadata.media_uri.as_ref()
-                    );
-                    assert_eq!(
-                        node_data.metadata.media_hash.into_bytes(),
-                        scanner_data.metadata.media_hash
-                    );
-                    assert_eq!(
-                        node_data.metadata.creator.map(|c| c.into_bytes()),
-                        scanner_data.metadata.creator.map(|c| c.encode())
-                    );
-                }
-            }
-        }
     }
 }
