+++ common/src/chain/tokens/rpc.rs
@@ -163,6 +163,12 @@ pub enum RPCIsTokenFrozen {
     Yes(RPCIsTokenUnfreezable),
 }
 
+impl From<IsTokenFrozen> for RPCIsTokenFrozen {
+    fn from(value: IsTokenFrozen) -> Self {
+        Self::new(value)
+    }
+}
+
 impl RPCIsTokenFrozen {
     pub fn new(frozen: IsTokenFrozen) -> Self {
         match frozen {
@@ -172,7 +178,7 @@ impl RPCIsTokenFrozen {
     }
 }
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, HasValueHint)]
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, HasValueHint)]
 pub struct RPCFungibleTokenInfo {
     // TODO: Add the controller public key to issuance data - https://github.com/mintlayer/mintlayer-core/issues/401
     pub token_id: TokenId,
@@ -237,7 +243,9 @@ impl RPCNonFungibleTokenInfo {
     }
 }
 
-#[derive(Debug, Clone, Encode, Decode, serde::Serialize, serde::Deserialize, HasValueHint)]
+#[derive(
+    Debug, Clone, PartialEq, Eq, Encode, Decode, serde::Serialize, serde::Deserialize, HasValueHint,
+)]
 pub struct RPCTokenCreator(Vec<u8>);
 
 impl RPCTokenCreator {
@@ -253,7 +261,7 @@ impl From<&TokenCreator> for RPCTokenCreator {
     }
 }
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, HasValueHint)]
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, HasValueHint)]
 pub struct RPCNonFungibleTokenMetadata {
     pub creator: Option<RPCTokenCreator>,
     pub name: RpcString,
